+++ src/test/java/org/folio/linked/data/e2e/mappings/work/hub/HubReferenceIT.java
@@ -20,9 +20,7 @@ class HubReferenceIT extends PostResourceIT {
             "http://bibfra.me/vocab/library/title": [
               {
                 "http://bibfra.me/vocab/library/Title": {
-                  "http://bibfra.me/vocab/library/mainTitle": [
-                    "%s"
-                  ]
+                  "http://bibfra.me/vocab/library/mainTitle": [ "%s" ]
                 }
               }
             ],
@@ -30,12 +28,15 @@ class HubReferenceIT extends PostResourceIT {
               {
                 "_relation": "http://bibfra.me/vocab/lite/expressionOf",
                 "_hub": {
-                  "http://bibfra.me/vocab/lite/label": [
-                    "hub label"
-                  ],
-                  "http://bibfra.me/vocab/lite/link": [
-                    "hub link"
-                  ]
+                  "http://bibfra.me/vocab/lite/label": [ "hub label 1" ],
+                  "http://bibfra.me/vocab/lite/link": [ "hub link 1" ]
+                }
+              },
+              {
+                "_relation": "http://bibfra.me/vocab/relation/relatedTo",
+                "_hub": {
+                  "http://bibfra.me/vocab/lite/label": [ "hub label 2" ],
+                  "http://bibfra.me/vocab/lite/link": [ "hub link 2" ]
                 }
               }
             ]
@@ -47,24 +48,34 @@ class HubReferenceIT extends PostResourceIT {
 
   @Override
   protected void validateGraph(Resource work) {
-    var hub = getFirstOutgoingResource(work, "http://bibfra.me/vocab/lite/expressionOf");
-    validateResourceType(hub, "http://bibfra.me/vocab/lite/Hub");
+    assertThat(work.getOutgoingEdges()).hasSize(3);
 
-    assertThat(hub.getId()).isEqualTo(1407893064845187544L);
-    assertThat(hub.getLabel()).isEqualTo("hub label");
-    assertThat(getProperty(hub, "http://bibfra.me/vocab/lite/label")).isEqualTo("hub label");
-    assertThat(getProperty(hub, "http://bibfra.me/vocab/lite/link")).isEqualTo("hub link");
+    var expressionOfHub = getFirstOutgoingResource(work, "http://bibfra.me/vocab/lite/expressionOf");
+    validateResourceType(expressionOfHub, "http://bibfra.me/vocab/lite/Hub");
+    assertThat(expressionOfHub.getId()).isEqualTo(-2469667353235297183L);
+    assertThat(expressionOfHub.getLabel()).isEqualTo("hub label 1");
+    assertThat(getProperty(expressionOfHub, "http://bibfra.me/vocab/lite/label")).isEqualTo("hub label 1");
+    assertThat(getProperty(expressionOfHub, "http://bibfra.me/vocab/lite/link")).isEqualTo("hub link 1");
+
+    var relatedToHub = getFirstOutgoingResource(work, "http://bibfra.me/vocab/relation/relatedTo");
+    validateResourceType(relatedToHub, "http://bibfra.me/vocab/lite/Hub");
+    assertThat(relatedToHub.getId()).isEqualTo(8331756295276184518L);
+    assertThat(relatedToHub.getLabel()).isEqualTo("hub label 2");
+    assertThat(getProperty(relatedToHub, "http://bibfra.me/vocab/lite/label")).isEqualTo("hub label 2");
+    assertThat(getProperty(relatedToHub, "http://bibfra.me/vocab/lite/link")).isEqualTo("hub link 2");
   }
 
   @Override
   @SneakyThrows
   protected void validateApiResponse(ResultActions apiResponse) {
     var workPath = "$.resource['http://bibfra.me/vocab/lite/Work']";
+    var relatedToPath = workPath + "._hubs[?(@._relation=='http://bibfra.me/vocab/relation/relatedTo')]._hub";
+    var expressionOfPath = workPath + "._hubs[?(@._relation=='http://bibfra.me/vocab/lite/expressionOf')]._hub";
 
     apiResponse
-      .andExpect(jsonPath(workPath + "['_hubs'][0]['_relation']").value("http://bibfra.me/vocab/lite/expressionOf"))
-      .andExpect(jsonPath(workPath + "['_hubs'][0]['_hub']['id']").value("1407893064845187544"))
-      .andExpect(jsonPath(workPath + "['_hubs'][0]['_hub']['http://bibfra.me/vocab/lite/label'][0]").value("hub label"))
-      .andExpect(jsonPath(workPath + "['_hubs'][0]['_hub']['http://bibfra.me/vocab/lite/link'][0]").value("hub link"));
+      .andExpect(jsonPath(expressionOfPath + "['http://bibfra.me/vocab/lite/label'][0]").value("hub label 1"))
+      .andExpect(jsonPath(expressionOfPath + "['http://bibfra.me/vocab/lite/link'][0]").value("hub link 1"))
+      .andExpect(jsonPath(relatedToPath + "['http://bibfra.me/vocab/lite/label'][0]").value("hub label 2"))
+      .andExpect(jsonPath(relatedToPath + "['http://bibfra.me/vocab/lite/link'][0]").value("hub link 2"));
   }
 }